pyproject.toml
[project]
name = "smartsummary"
version = "0.1.0"
description = "Telegram userbot for chat monitoring and AI summarization"
requires-python = ">=3.11"
dependencies = [
    "telethon>=1.37,<2.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.7",
    "anthropic>=0.45",
    "openai>=1.0",
    "apscheduler>=3.10,<4.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
]
